넉백, 플레이어, 몬스터 관련 일부 수정

+++ Assets/Scripts/PassiveSystem.cs
@@ -9,7 +9,7 @@ public class PassiveSystem : MonoBehaviour
     [SerializeField] private Battle battle;
     [SerializeField] private PlayerStatus status;
 
-    public float[] passiveRate;
+    public int[] passiveRate;
 
     [Header("현재 상태")]
     public bool isGetBarrier;
